Fixed bugs on parser.py, added instructions and colors, and renamed elements.
Elements renamed to improve readability are as follows:
    functions to functionDecs
    function to functionDec
    assign_or_funccall to assignOrFunccall
    funccall to functionCall
    colorconstant to colorConstant

Fixed error in functionCall: changed "params" to "args".
Fixed error in "assignment", "condition", "loop" and "print", where the semicolon ";" was referenced twice.

Added instructions format to draw "circle", "square" and "arc".
Added more colors to colorConstant in order to reflect the changes in lexer.py.

See documentation for the graphical representation of the syntax diagrams.